HZN-1342: Move (Alarm)StatisticsService to opennms-dao-api
