+++ features/distributed/dao-test/src/main/java/org/opennms/netmgt/dao/DaoServiceTester.java
@@ -53,7 +53,7 @@ import org.opennms.netmgt.dao.api.NodeLabel;
 import org.opennms.netmgt.dao.api.OnmsDao;
 import org.opennms.netmgt.dao.api.SessionFactoryWrapper;
 import org.opennms.netmgt.dao.api.TopologyDao;
-import org.opennms.netmgt.dao.stats.StatisticsService;
+import org.opennms.netmgt.dao.api.StatisticsService;
 import org.opennms.netmgt.filter.api.FilterDao;
 import org.opennms.netmgt.model.FilterManager;
 import org.opennms.netmgt.model.OnmsNode;
