+++ opennms-dao-api/src/main/java/org/opennms/netmgt/dao/api/AlarmStatisticsService.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.dao.stats;
+package org.opennms.netmgt.dao.api;
 
 import org.opennms.core.criteria.Criteria;
 import org.opennms.netmgt.model.OnmsAlarm;
