+++ opennms-dao/src/main/java/org/opennms/netmgt/dao/stats/DefaultAlarmStatisticsService.java
@@ -32,6 +32,7 @@ import java.util.List;
 
 import org.opennms.core.criteria.Criteria;
 import org.opennms.netmgt.dao.api.AlarmDao;
+import org.opennms.netmgt.dao.api.AlarmStatisticsService;
 import org.opennms.netmgt.model.OnmsAlarm;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
