+++ opennms-webapp-rest/src/main/java/org/opennms/web/rest/v1/AlarmStatsRestService.java
@@ -51,7 +51,7 @@ import org.opennms.core.criteria.Alias.JoinType;
 import org.opennms.core.criteria.Criteria;
 import org.opennms.core.criteria.CriteriaBuilder;
 import org.opennms.core.criteria.Fetch.FetchType;
-import org.opennms.netmgt.dao.stats.AlarmStatisticsService;
+import org.opennms.netmgt.dao.api.AlarmStatisticsService;
 import org.opennms.netmgt.model.OnmsAlarm;
 import org.opennms.netmgt.model.OnmsSeverity;
 import org.slf4j.Logger;
